backend/requirements.txt
fastapi
uvicorn
python-multipart
sentence-transformers
faiss-cpu
PyMuPDF
python-docx
python-dotenv
google-generativeai
